+++ contentpack-editor/main/src/com/pixurvival/contentPackEditor/component/valueComponent/AngleInput.java
@@ -8,12 +8,12 @@ import com.pixurvival.core.util.MathUtils;
 
 import lombok.SneakyThrows;
 
-public class AngleInput extends NumberInput<Double> {
+public class AngleInput extends NumberInput<Float> {
 
 	private static final long serialVersionUID = 1L;
 
-	private static final double DEG_TO_RAD = Math.PI / 180;
-	private static final double RAD_TO_DEG = 180 / Math.PI;
+	private static final float DEG_TO_RAD = (float) Math.PI / 180;
+	private static final float RAD_TO_DEG = 180 / (float) Math.PI;
 	private static final DecimalFormat DECIMAL_FORMAT = new DecimalFormat("0.##", new DecimalFormatSymbols(Locale.US));
 
 	public AngleInput() {
@@ -22,16 +22,16 @@ public class AngleInput extends NumberInput<Double> {
 
 	@Override
 	@SneakyThrows
-	protected Double parse(String text) {
+	protected Float parse(String text) {
 		if (text.matches("\\-?\\d+(\\.\\d+)?")) {
-			return MathUtils.normalizeAngle(DECIMAL_FORMAT.parse(text).doubleValue() * DEG_TO_RAD);
+			return MathUtils.normalizeAngle(DECIMAL_FORMAT.parse(text).floatValue() * DEG_TO_RAD);
 		} else {
 			return null;
 		}
 	}
 
 	@Override
-	protected String format(Double value) {
+	protected String format(Float value) {
 		return DECIMAL_FORMAT.format(value * RAD_TO_DEG);
 	}
