+++ core/main/src/com/pixurvival/core/Direction.java
@@ -12,15 +12,15 @@ import lombok.Getter;
 public enum Direction {
 
 	EAST(0),
-	NORTH_EAST(Math.PI / 4.0),
-	NORTH(Math.PI / 2.0),
-	NORTH_WEST(Math.PI * 3.0 / 4.0),
-	WEST(Math.PI),
-	SOUTH_WEST(-Math.PI * 3.0 / 4.0),
-	SOUTH(-Math.PI / 2.0),
-	SOUTH_EAST(-Math.PI / 4.0);
+	NORTH_EAST((float) Math.PI / 4f),
+	NORTH((float) Math.PI / 2f),
+	NORTH_WEST((float) Math.PI * 3f / 4f),
+	WEST((float) Math.PI),
+	SOUTH_WEST(-(float) Math.PI * 3f / 4f),
+	SOUTH(-(float) Math.PI / 2f),
+	SOUTH_EAST(-(float) Math.PI / 4f);
 
-	private double angle;
+	private float angle;
 
 	/**
 	 * Returns the closest cardinal direction of the given angle.
@@ -31,10 +31,10 @@ public enum Direction {
 	 *         ambiguous angles (e.g. Pi/4), The closest cardinal to south is
 	 *         given.
 	 */
-	public static Direction closestCardinalDirection(double angle) {
-		if (angle <= Math.PI / 4 && angle > -Math.PI / 4) {
+	public static Direction closestCardinalDirection(float angle) {
+		if (angle <= (float) Math.PI / 4 && angle > -(float) Math.PI / 4) {
 			return EAST;
-		} else if (angle >= 3.0 / 4.0 * Math.PI || angle < -3.0 / 4.0 * Math.PI) {
+		} else if (angle >= 3f / 4f * (float) Math.PI || angle < -3f / 4f * (float) Math.PI) {
 			return WEST;
 		} else if (angle > 0) {
 			return NORTH;
