+++ core/main/src/com/pixurvival/core/World.java
@@ -162,7 +162,7 @@ public class World extends PluginHolder<World> implements ChatSender {
 		listeners.add(listener);
 	}
 
-	public synchronized void update(double deltaTimeMillis) {
+	public synchronized void update(float deltaTimeMillis) {
 		if (gameEnded) {
 			return;
 		}
@@ -188,8 +188,8 @@ public class World extends PluginHolder<World> implements ChatSender {
 	}
 
 	/**
-	 * Called after all players are added in the EntityPool and Teams are sets. This
-	 * will place players and set the map limit if present.
+	 * Called after all players are added in the EntityPool and Teams are sets.
+	 * This will place players and set the map limit if present.
 	 */
 	public void initializeGame() {
 		entityPool.flushNewEntities();
